baseline/requirements.txt
numpy==1.24.4
scipy
pandas==1.1.5
tensorboard==2.4.1
flake8==3.3.0
pep8-naming==0.4.1
scikit-learn==0.24.0
pytest==3.2.1

tabulate==0.8.9
hyperopt==0.2.5
tqdm==4.60.0
geopy==2.0.0
dgl==0.6.1
fastdtw==0.3.4
networkx==2.5.1
gensim
torchtext==0.6.0
nltk==3.2.4
statsmodels
tensorboardx==2.2
aioredis==1.3.1
aiohttp==3.7.4
idna==2.10
torchdiffeq==0.2.3
# infomap  # Windows需要C++编译器，可选依赖，先注释掉
einops
torch-geometric  # 使用连字符格式